Sentence editing
Section 3.3, under "Odyssey Court Software".
The first sentence starts with "since 2016" which implies that the crisis is still going on, thus the correct verb should be "have been having".
In the same sentence, near the end, "recording and disseminating" is the correct pair since both have to be in the "-ing" form (either both or neither).
Also the url in the "Details" part is too long for the width of an A4 paper. I don't know how you would want to fix that. (I suggest \href with a hyperlink over the word "Details", however this would imply change for all other links (I can fix those).