Sentence editing in section 3.5.1 - OS part
Replaced "critical" with "dangerous", because "critical" means important (or "of high value"). As such it leads to imply that vulnerabilities are important, which is misleading. Vulnerabilities are dangerous; acknowledging and fixing them is important. 

"so that" is replaced by "therefore". "So that" implies that there is a reason for only using a few OS, and that reason is to have many users that share similar problems. This is also misleading.

Two other sentences were edited as shown by preview.